remove some logging and undo containerlogger move
