+++ pkg/worker/image.go
@@ -23,6 +23,7 @@ import (
 	"github.com/opencontainers/umoci/oci/casext"
 	"github.com/opencontainers/umoci/oci/layer"
 	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
 
 	"github.com/beam-cloud/beta9/pkg/abstractions/image"
 	common "github.com/beam-cloud/beta9/pkg/common"
@@ -78,6 +79,13 @@ type ImageClient struct {
 	config             types.AppConfig
 	workerId           string
 	workerRepo         repository.WorkerRepository
+}
+
+type logger struct {
+	logFile     *os.File
+	logger      *logrus.Logger
+	fields      logrus.Fields
+	containerId string
 }
 
 func NewImageClient(config types.AppConfig, workerId string, workerRepo repository.WorkerRepository) (*ImageClient, error) {
@@ -138,7 +146,7 @@ func (c *ImageClient) PullLazy(request *types.ContainerRequest) error {
 
 	isBuildContainer := strings.HasPrefix(request.ContainerId, types.BuildContainerPrefix)
 
-	f, err := NewLogger(request)
+	f, err := newLogger(request)
 	if err != nil {
 		return err
 	}
@@ -408,4 +416,30 @@ func (c *ImageClient) Archive(ctx context.Context, bundlePath string, imageId st
 
 	log.Printf("Image <%v> push took %v\n", imageId, time.Since(startTime))
 	return nil
+}
+
+func (l *logger) Log(format string, args ...any) {
+	log.Print(fmt.Sprintf("<%s> - ", l.containerId) + fmt.Sprintf(format, args...))
+	l.logger.WithFields(l.fields).Infof(format, args...)
+}
+
+func newLogger(request *types.ContainerRequest) (*logger, error) {
+	logFile, err := openLogFile(request.ContainerId)
+	if err != nil {
+		return nil, err
+	}
+
+	f := logrus.New()
+	f.SetOutput(logFile)
+	f.SetFormatter(&logrus.JSONFormatter{TimestampFormat: time.RFC3339Nano})
+
+	return &logger{
+		logFile: logFile,
+		logger:  f,
+		fields: logrus.Fields{
+			"container_id": request.ContainerId,
+			"stub_id":      request.StubId,
+		},
+		containerId: request.ContainerId,
+	}, nil
 }
